Install packages and make file structure to begin testing
